pr comments rename cromwell stuff to cbas make permutation of test inputs more consistent
